Extract Series classes
